+++ src/Xbehave.Sdk.Net40/Step.cs
@@ -7,24 +7,24 @@ namespace Xbehave.Sdk
     using System;
     using System.Collections.Generic;
     using System.Diagnostics.CodeAnalysis;
+    using System.Linq;
     using System.Reflection;
+    using System.Threading;
+    using Xunit.Sdk;
 
+    /// <summary>
+    /// Provides the implementation to execute each step.
+    /// </summary>
     [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Step", Justification = "By design.")]
-    public abstract class Step
+    public abstract partial class Step
     {
-        protected readonly MethodInfo MethodInfo;
-        protected readonly object Target;
         protected readonly List<Action> Teardowns = new List<Action>();
         private readonly string name;
         private readonly object stepType;
 
-        public Step(string name, MethodInfo methodInfo, object target, object stepType)
+        public Step(string name, object stepType)
         {
-            Guard.AgainstNullArgument("methodInfo", methodInfo);
-
             this.name = name;
-            this.MethodInfo = methodInfo;
-            this.Target = target;
             this.stepType = stepType;
         }
 
@@ -52,11 +52,84 @@ namespace Xbehave.Sdk
             }
         }
 
-        public abstract void Execute();
-
-        protected object ExecuteMethodInfo()
+        public void Execute()
         {
-            return this.MethodInfo.Invoke(this.Target, null);
+            IEnumerable<Action> teardowns = Enumerable.Empty<Action>();
+
+            try
+            {
+                Exception ex = null;
+
+                ManualResetEvent @event = new ManualResetEvent(false);
+
+                ThreadPool.QueueUserWorkItem(o =>
+                {
+                    var oldSyncContext = SynchronizationContext.Current;
+
+                    try
+                    {
+                        this.SetupSynchronizationContext();
+
+                        this.ExecuteBody();
+
+                        ex = this.WaitForCompletion();
+                    }
+                    catch (TargetInvocationException targetEx)
+                    {
+                        ex = targetEx.InnerException;
+                    }
+                    catch (Exception e)
+                    {
+                        ex = e;
+                    }
+                    finally
+                    {
+                        teardowns = CurrentScenario.ExtractTeardowns();
+                        this.TeardownSynchronizationContext(oldSyncContext);
+
+                        @event.Set();
+                    }
+                });
+
+                if (this.MillisecondsTimeout > 0)
+                {
+                    // NOTE: we do not call the WaitOne(int) overload because it wasn't introduced until .NET 3.5 SP1 and we want to support pre-SP1
+                    if (!@event.WaitOne(this.MillisecondsTimeout, false))
+                    {
+                        throw new Xunit.Sdk.TimeoutException(this.MillisecondsTimeout);
+                    }
+                }
+                else
+                {
+                    @event.WaitOne();
+                }
+
+                if (ex != null)
+                {
+                    ExceptionUtility.RethrowWithNoStackTraceLoss(ex);
+                }
+            }
+            finally
+            {
+                foreach (var teardown in teardowns)
+                {
+                    CurrentScenario.AddTeardown(teardown);
+                }
+
+                this.Teardowns.ForEach(CurrentScenario.AddTeardown);
+            }
+        }
+
+        protected abstract void ExecuteBody();
+
+        protected abstract Exception WaitForCompletion();
+
+        protected virtual void SetupSynchronizationContext()
+        {
+        }
+
+        protected virtual void TeardownSynchronizationContext(SynchronizationContext oldSyncContext)
+        {
         }
     }
 }
