+++ src/Xbehave.Sdk.Net40/Xbehave.Sdk.Net40.csproj
@@ -62,7 +62,7 @@
       <Link>Properties\CommonAssemblyInfo.cs</Link>
     </Compile>
     <Compile Include="Argument.cs" />
-    <Compile Include="AsyncTestSyncContext.cs" />
+    <Compile Include="Step.AsyncTestSyncContext.cs" />
     <Compile Include="ContextCommand.cs" />
     <Compile Include="BackgroundCommand.cs" />
     <Compile Include="Guard.cs" />
