+++ src/Xbehave.Sdk.Net45/Xbehave.Sdk.Net45.csproj
@@ -65,9 +65,6 @@
     <Compile Include="..\Xbehave.Sdk.Net40\Argument.cs">
       <Link>Argument.cs</Link>
     </Compile>
-    <Compile Include="..\Xbehave.Sdk.Net40\AsyncTestSyncContext.cs">
-      <Link>AsyncTestSyncContext.cs</Link>
-    </Compile>
     <Compile Include="..\Xbehave.Sdk.Net40\BackgroundCommand.cs">
       <Link>BackgroundCommand.cs</Link>
     </Compile>
@@ -100,6 +97,9 @@
     </Compile>
     <Compile Include="..\Xbehave.Sdk.Net40\ReplayCommand.cs">
       <Link>ReplayCommand.cs</Link>
+    </Compile>
+    <Compile Include="..\Xbehave.Sdk.Net40\Step.AsyncTestSyncContext.cs">
+      <Link>Step.AsyncTestSyncContext.cs</Link>
     </Compile>
     <Compile Include="..\Xbehave.Sdk.Net40\Step.cs">
       <Link>Step.cs</Link>
